Move Docker tests into more appropriate location
Most of the tests in `pulp_smash.tests.docker.cli.test_sync_publish` use
Pulp's CLI and API interfaces. Rewrite the tests to use just Pulp's API,
and move them to `pulp_smash.tests.docker.api_v2.test_sync_publish`.
