+++ docs/api.rst
@@ -24,6 +24,7 @@ developers, not a gospel.
     api/pulp_smash.tests.docker.api_v2.test_crud
     api/pulp_smash.tests.docker.api_v2.test_duplicate_uploads
     api/pulp_smash.tests.docker.api_v2.test_sync
+    api/pulp_smash.tests.docker.api_v2.test_sync_publish
     api/pulp_smash.tests.docker.api_v2.test_tags
     api/pulp_smash.tests.docker.api_v2.utils
     api/pulp_smash.tests.docker.cli
