+++ docs/api/pulp_smash.tests.docker.api_v2.test_sync_publish.rst
@@ -1,0 +1,6 @@
+`pulp_smash.tests.docker.api_v2.test_sync_publish`
+==================================================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.tests.docker.api_v2.test_sync_publish`
+
+.. automodule:: pulp_smash.tests.docker.api_v2.test_sync_publish
